requirements.txt
flask==3.0.0
werkzeug==3.0.1
waitress==3.0.0
vercel==0.3.5
python-dotenv==1.0.0